Make volume parameter a size_t

+++ arb_dirac_eigen/lattice.h
@@ -94,7 +94,7 @@ EXTERN  int Kiters ; /* Kalkreuter iterations */
 
 /* The following are global scalars */
 EXTERN	int nx,ny,nz,nt;	/* lattice dimensions */
-EXTERN  int volume;		/* volume of lattice = nx*ny*nz*nt */
+EXTERN  size_t volume;		/* volume of lattice = nx*ny*nz*nt */
 #ifdef RANDOM
 EXTERN  int iseed;              /* random number seed */
 #endif
