+++ rcorr/lattice.h
@@ -27,7 +27,7 @@ typedef struct {
 
 EXTERN	int nx,ny,nz,nt;	/* lattice dimensions */
 EXTERN  int iseed;              /* Not used in this application */
-EXTERN  int volume;		/* volume of lattice = nx*ny*nz*nt */
+EXTERN  size_t volume;		/* volume of lattice = nx*ny*nz*nt */
 
 #ifdef FIX_NODE_GEOM
 EXTERN  int node_geometry[4];  /* Specifies fixed "nsquares" (i.e. 4D
